added two more deleted rules
